requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.58.0
imageio==2.37.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycryptodome==3.22.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
pyzmq==26.4.0
# Editable install with no version control (rsa_dwt_project==0.1)
-e d:\vscode\cryptography\project
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
tifffile==2025.5.10
tornado==6.5
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
